Change log drops to use number of logs(float) instead of voxels(int)

+++ src/main/java/com/ferreusveritas/dynamictrees/api/treedata/IDropCreator.java
@@ -73,6 +73,6 @@ public interface IDropCreator {
 	 * @param volume
 	 * @return
 	 */
-	List<ItemStack> getLogsDrop(World world, Species species, BlockPos breakPos, Random random, List<ItemStack> dropList, int volume);
+	List<ItemStack> getLogsDrop(World world, Species species, BlockPos breakPos, Random random, List<ItemStack> dropList, float volume);
 	
 }
